Start defining GameState2
This is kind of an evolution of TurnData. I don't think that the
gamestate of setup/inplay/finish is as useful as "Where we are now" at
any given point.

But then instead of trying to make a TurnData so fine grained with
phases just have a single `Turn` object that holds all of that. To
build a `Turn` we'll probably still have some phase element.